# upravljanjevremenom/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Time Manager API",
        "description": "REST API sistem za upravljanje vremenom. \nOmogućava korisnicima upravljanje zadacima, beleškama i obaveštenjima, \npraćenje statusa zadataka, kao i pregled statistike rada korisnika. \nSistem podržava uloge (admin, regular i premium korisnik).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/statistike": {
            "get": {
                "tags": [
                    "Statistika"
                ],
                "summary": "Lista svih statistika",
                "operationId": "29b7d83ef9fdedc6b8b3aa40da4cdc63",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista statistika"
                    }
                }
            },
            "post": {
                "tags": [
                    "Statistika"
                ],
                "summary": "Kreiranje nove statistike",
                "operationId": "d553cf89f7a619c9390211ddee2e3d9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "broj_odradjenih_zadataka",
                                    "ukupan_broj_zadataka",
                                    "procenat_uspesnosti",
                                    "korisnik_id"
                                ],
                                "properties": {
                                    "broj_odradjenih_zadataka": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "ukupan_broj_zadataka": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "procenat_uspesnosti": {
                                        "type": "number",
                                        "example": 80
                                    },
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Statistika uspešno kreirana"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/statistike/{id}": {
            "get": {
                "tags": [
                    "Statistika"
                ],
                "summary": "Prikaz jedne statistike",
                "operationId": "02182d554f027145730902ef407b87f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistika pronađena"
                    },
                    "404": {
                        "description": "Statistika nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Statistika"
                ],
                "summary": "Izmena statistike",
                "operationId": "14764ada0c968a695e83c5acd708730f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "broj_odradjenih_zadataka": {
                                        "type": "integer",
                                        "example": 9
                                    },
                                    "ukupan_broj_zadataka": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "procenat_uspesnosti": {
                                        "type": "number",
                                        "example": 75
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statistika uspešno izmenjena"
                    },
                    "404": {
                        "description": "Statistika nije pronađena"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Statistika"
                ],
                "summary": "Brisanje statistike",
                "operationId": "64d2631ed289e2b4bba47121b824e30d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistika uspešno obrisana"
                    },
                    "404": {
                        "description": "Statistika nije pronađena"
                    }
                }
            }
        },
...
    },
...
}